# make tail a btree of jumps