... for fun.

This takes 128 tail from a spread of 9 to 14ms to 10 to 12.

... every nanosecond matters

So, while this is slower for the tail length 1 and 2 cases, it's about
equal or faster for all others. And by slower, I mean .8ns.

benchmark                       old ns/op     new ns/op     delta
BenchmarkBase32-32              5.78          5.73          -0.87%
BenchmarkLoop32-32              6.79          6.76          -0.44%
BenchmarkTail32_1-32            6.40          6.38          -0.31%
BenchmarkTail32_2-32            6.50          6.46          -0.62%
BenchmarkTail32_3-32            6.82          6.95          +1.91%
BenchmarkBase128-32             7.98          7.94          -0.50%
BenchmarkLoop128-32             12.1          11.9          -1.65%
BenchmarkTail128_1-32           9.83          10.6          +7.83%
BenchmarkTail128_2-32           9.68          10.3          +6.40%
BenchmarkTail128_3-32           10.4          10.2          -1.92%
BenchmarkTail128_4-32           10.2          10.5          +2.94%
BenchmarkTail128_5-32           11.1          11.4          +2.70%
BenchmarkTail128_6-32           11.5          11.1          -3.48%
BenchmarkTail128_7-32           12.1          10.5          -13.22%
BenchmarkTail128_8-32           11.5          10.5          -8.70%
BenchmarkTail128_9-32           12.8          11.9          -7.03%
BenchmarkTail128_10-32          12.9          11.4          -11.63%
BenchmarkTail128_11-32          13.3          10.9          -18.05%
BenchmarkTail128_12-32          13.3          12.0          -9.77%
BenchmarkTail128_13-32          14.0          12.6          -10.00%
BenchmarkTail128_14-32          14.4          12.3          -14.58%
BenchmarkTail128_15-32          14.4          12.4          -13.89%